chore: move the delivery of the GTFS files to the feed app
